/stats command Moved multiple functions to commandLibrary for readability
